refactor(src): split code in multiple files
