+++ src/errors.ts
@@ -1,0 +1,9 @@
+export class ETOOBigError extends Error {
+  code: string;
+
+  constructor() {
+    super('maximum file length exceeded');
+
+    this.code = 'ETOOBIG';
+  }
+}
